requirements.txt

streamlit
qrcode[pil]
pillow
